Use decimal to keep money

+++ storage/psql/db/migrations/000002_create_transactions_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE transactions(
     id serial PRIMARY KEY,
     user_id integer REFERENCES users ON DELETE CASCADE,
     order_id VARCHAR(255) NOT NULL,
-    amount integer NOT NULL,
+    amount decimal NOT NULL,
     created_at timestamp
 );
 CREATE INDEX ON transactions (user_id);
